week 2/week2-agent-api-langgraph/Day 7/requirements.txt
fastapi
uvicorn
openai>=1.9.0
langchain>=0.2.9
langchain-openai
langchain-community
langchain-chroma
langgraph>=0.0.40
langchain-core
chromadb
python-dotenv
PyPDF2
tiktoken